# Take advantage of sync.Pool in worktree.Status()