--- a/plumbing/format/objfile/reader.go
+++ b/plumbing/format/objfile/reader.go
@@ -21,11 +21,10 @@ var (
 // Reader implements io.ReadCloser. Close should be called when finished with
 // the Reader. Close will not close the underlying io.Reader.
 type Reader struct {
-	multi   io.Reader
-	zlib    io.Reader
-	zlibref sync.ZLibReader
-	hasher  plumbing.Hasher
-	closed  bool
+	multi  io.Reader
+	zlib   *sync.ZLibReader
+	hasher plumbing.Hasher
+	closed bool
 }
 
 // NewReader returns a new Reader reading from r.
@@ -35,10 +34,7 @@ func NewReader(r io.Reader) (*Reader, error) {
 		return nil, packfile.ErrZLib.AddDetails(err.Error())
 	}
 
-	return &Reader{
-		zlib:    zlib.Reader,
-		zlibref: zlib,
-	}, nil
+	return &Reader{zlib: zlib}, nil
 }
 
 // Header reads the type and the size of object, and prepares the reader for read
@@ -118,6 +114,7 @@ func (r *Reader) Close() error {
 		return nil
 	}
 	r.closed = true
-	sync.PutZlibReader(r.zlibref)
-	return nil
+
+	defer sync.PutZlibReader(r.zlib)
+	return r.zlib.Close()
 }