--- a/utils/sync/zlib.go
+++ b/utils/sync/zlib.go
@@ -11,7 +11,11 @@ var (
 	zlibInitBytes = []byte{0x78, 0x9c, 0x01, 0x00, 0x00, 0xff, 0xff, 0x00, 0x00, 0x00, 0x01}
 	zlibReader    = sync.Pool{
 		New: func() interface{} {
-			return NewZlibReader(nil)
+			r, _ := zlib.NewReader(bytes.NewReader(zlibInitBytes))
+			return &ZLibReader{
+				reader: r.(zlibReadCloser),
+				dict:   nil,
+			}
 		},
 	}
 	zlibWriter = sync.Pool{
@@ -26,25 +30,17 @@ type zlibReadCloser interface {
 	zlib.Resetter
 }
 
-func NewZlibReader(dict *[]byte) ZLibReader {
-	r, _ := zlib.NewReader(bytes.NewReader(zlibInitBytes))
-	return ZLibReader{
-		Reader: r.(zlibReadCloser),
-		dict:   dict,
-	}
-}
-
 type ZLibReader struct {
 	dict   *[]byte
-	Reader zlibReadCloser
+	reader zlibReadCloser
 }
 
-func (z ZLibReader) Reset(r io.Reader) error {
-	var dict []byte
-	if z.dict != nil {
-		dict = *z.dict
-	}
-	return z.Reader.Reset(r, dict)
+func (r *ZLibReader) Read(p []byte) (int, error) {
+	return r.reader.Read(p)
+}
+
+func (r *ZLibReader) Close() error {
+	return r.reader.Close()
 }
 
 // GetZlibReader returns a ZLibReader that is managed by a sync.Pool.
@@ -53,19 +49,18 @@ func (z ZLibReader) Reset(r io.Reader) error {
 //
 // After use, the ZLibReader should be put back into the sync.Pool
 // by calling PutZlibReader.
-func GetZlibReader(r io.Reader) (ZLibReader, error) {
-	z := zlibReader.Get().(ZLibReader)
+func GetZlibReader(r io.Reader) (*ZLibReader, error) {
+	z := zlibReader.Get().(*ZLibReader)
 	z.dict = GetByteSlice()
 
-	err := z.Reader.Reset(r, *z.dict)
+	err := z.reader.Reset(r, *z.dict)
 
 	return z, err
 }
 
-// PutZlibReader puts z back into its sync.Pool, first closing the reader.
+// PutZlibReader puts z back into its sync.Pool.
 // The Byte slice dictionary is also put back into its sync.Pool.
-func PutZlibReader(z ZLibReader) {
-	z.Reader.Close()
+func PutZlibReader(z *ZLibReader) {
 	PutByteSlice(z.dict)
 	zlibReader.Put(z)
 }