Depends on #1653. I'll rebase after it gets merged.

[Download pprof diff](https://github.com/user-attachments/files/22248592/profile.pb.gz)

worktree.Status() is now 21% faster and allocates 93% less memory.
```
goos: darwin
goarch: arm64
pkg: github.com/go-git/go-git/v6
cpu: Apple M2
                 │   old.out   │              new4.out               │
                 │   sec/op    │   sec/op     vs base                │
WorktreeStatus-8   494.3µ ± 5%   390.1µ ± 1%  -21.07% (p=0.000 n=10)

                 │    old.out     │               new4.out               │
                 │      B/op      │     B/op      vs base                │
WorktreeStatus-8   1354.38Ki ± 0%   90.50Ki ± 0%  -93.32% (p=0.000 n=10)

                 │   old.out   │              new4.out              │
                 │  allocs/op  │  allocs/op   vs base               │
WorktreeStatus-8   2.094k ± 0%   1.897k ± 0%  -9.41% (p=0.000 n=10)
```

NOTE: 
Those tests in sync package are flaky. See 80520758eab82f587f40757c3b4db120e8e562c4.
You can test it using `go test ./utils/sync/... -count=10000`.